Split shell abandoned sync tests
